Move tasks of activity parser class to an endpoint
The bean running at every instance of opening server increases the server load a bit. So I have decided to move the task of that parsing class to an endpoint to improve the structure of the project. Also, by having an endpoint, if we were to change the activities.json file to a newer one from the activity bank, it would easily load. We will have a load button in the admin panel to reload activities from activities.json file.